Change the tkinter template
Switches to a much simpler example of a tkinter template, taking it directly from the tk docs at http://www.tkdocs.com/tutorial/firstexample.html#code. Adds an additional row for the shift value as an additional input as well. Changes all of the variables in the tkinter section to more appropriate names.